Refactor ChecksResults and ClusterDetails into separate components
