+++ assets/js/components/ChecksResults/ChecksResults.jsx
@@ -1,34 +1,33 @@
 import React, { useEffect, useState } from 'react';
-import { useNavigate, useParams } from 'react-router-dom';
+import { useParams } from 'react-router-dom';
 import { useSelector, useDispatch } from 'react-redux';
 
+import ReactMarkdown from 'react-markdown';
+import remarkGfm from 'remark-gfm';
+
+import classNames from 'classnames';
+
+import { EOS_ERROR } from 'eos-icons-react';
+
+import { getCluster } from '@state/selectors';
+
 import Modal from '@components/Modal';
+import NotificationBox from '@components/NotificationBox';
+import LoadingBox from '@components/LoadingBox';
 
-import { EOS_ERROR, EOS_SETTINGS, EOS_PLAY_CIRCLE } from 'eos-icons-react';
-
-import NotificationBox from '../NotificationBox';
-import LoadingBox from '../LoadingBox';
-
-import Button from '@components/Button';
 import BackButton from '@components/BackButton';
 import WarningBanner from '@components/Banners/WarningBanner';
 import { UNKNOWN_PROVIDER } from '@components/ClusterDetails/ClusterSettings';
 
-import { getCluster } from '@state/selectors';
-import TrentoLogo from '../../../static/trento-icon.png';
-import {
-  TriggerChecksExecutionRequest,
-  truncatedClusterNameClasses,
-} from './ClusterDetails';
-import { ExecutionIcon } from './ExecutionIcon';
 import { getClusterName } from '@components/ClusterLink';
-import ChecksResultFilters, {
-  useFilteredChecks,
-} from '@components/ClusterDetails/ChecksResultFilters';
-import classNames from 'classnames';
 
-import ReactMarkdown from 'react-markdown';
-import remarkGfm from 'remark-gfm';
+import { ExecutionIcon } from './ExecutionIcon';
+import ChecksResultFilters, { useFilteredChecks } from './ChecksResultFilters';
+import ChecksSelectionHints from './ChecksSelectionHints';
+
+const truncatedClusterNameClasses = classNames(
+  'font-bold truncate w-60 inline-block align-top'
+);
 
 const getHostname =
   (hosts = []) =>
@@ -58,7 +57,7 @@ const getResultIcon = (executionState, health) => {
   return <ExecutionIcon health={health} executionState={executionState} />;
 };
 
-export const ChecksResults = () => {
+const ChecksResults = () => {
   const [modalOpen, setModalOpen] = useState(false);
   const [selectedCheck, setSelectedCheck] = useState('');
   const dispatch = useDispatch();
@@ -117,7 +116,7 @@ export const ChecksResults = () => {
     );
   } else if (!hasAlreadyChecksResults) {
     pageContent = (
-      <HintForChecksSelection
+      <ChecksSelectionHints
         clusterId={clusterID}
         selectedChecks={cluster?.selected_checks}
       />
@@ -257,52 +256,4 @@ export const ChecksResults = () => {
   );
 };
 
-const HintForChecksSelection = ({ clusterId, selectedChecks }) => {
-  const navigate = useNavigate();
-
-  const hasSelectedChecks = selectedChecks.length > 0;
-
-  return (
-    <div className="flex items-center justify-center py-5">
-      <div className="w-full rounded-lg bg-white dark:bg-gray-800 shadow-lg px-5 pt-5 pb-10 text-gray-800 dark:text-gray-50">
-        <div className="w-full pt-8 text-center pb-5 -mt-16 mx-auto">
-          <img
-            alt="profil"
-            src={TrentoLogo}
-            className="mx-auto object-cover h-20 w-20 "
-          />
-        </div>
-        <div className="w-full mb-10">
-          <p className="ttext-gray-600 dark:text-gray-100 text-center px-5">
-            {!hasSelectedChecks &&
-              'It looks like you have not configured any checks for the current cluster. Select your desired checks to be executed.'}
-            {hasSelectedChecks &&
-              'It looks like there is no recent execution for current cluster. Run your Check selection now!'}
-          </p>
-        </div>
-        <div className="w-full text-center">
-          {!hasSelectedChecks && (
-            <Button
-              className="bg-waterhole-blue mx-auto px-2 py-2 w-1/4 xs:w-full"
-              onClick={() => {
-                navigate(`/clusters/${clusterId}/settings`);
-              }}
-            >
-              <EOS_SETTINGS className="inline-block fill-white mr-1" />
-              Select Checks now!
-            </Button>
-          )}
-          {hasSelectedChecks && (
-            <TriggerChecksExecutionRequest
-              cssClasses="rounded relative w-1/4 ml-0.5 bg-waterhole-blue mx-auto px-2 py-2 w-1/4 xs:w-full text-base"
-              clusterId={clusterId}
-            >
-              <EOS_PLAY_CIRCLE className="inline-block fill-white" /> Start
-              Execution now
-            </TriggerChecksExecutionRequest>
-          )}
-        </div>
-      </div>
-    </div>
-  );
-};
+export default ChecksResults;
