+++ assets/js/components/ClusterDetails/index.js
@@ -1,7 +1,6 @@
 import ClusterDetails from './ClusterDetails';
 
 export { ClusterSettings } from './ClusterSettings';
-export { ChecksResults } from './ChecksResults';
 export { ConnectionSettings } from './ConnectionSettings';
 export { ExecutionIcon } from './ExecutionIcon';
