+++ assets/js/components/Tooltip/index.js
@@ -1,0 +1,3 @@
+import Tooltip from './Tooltip';
+
+export default Tooltip;
