+++ assets/js/components/TriggerChecksExecutionRequest/index.js
@@ -1,0 +1,3 @@
+import TriggerChecksExecutionRequest from './TriggerChecksExecutionRequest';
+
+export default TriggerChecksExecutionRequest;
