+++ assets/js/trento.jsx
@@ -15,7 +15,8 @@ import Home from '@components/Home';
 import AboutPage from '@components/AboutPage';
 import HostsList from '@components/HostsList';
 import ClustersList from '@components/ClustersList';
-import { ClusterSettings, ChecksResults } from '@components/ClusterDetails';
+import { ClusterSettings } from '@components/ClusterDetails';
+import ChecksResults from '@components/ChecksResults';
 import SapSystemsOverview from '@components/SapSystemsOverview';
 import HostDetails from '@components/HostDetails';
 import ClusterDetails from '@components/ClusterDetails';
